fix cuda support for cuda. add test that ensures Keras gradient return values.
