+++ tests/test_16_KerasModel.py
@@ -9,6 +9,8 @@ import kipoi
 from kipoi.pipeline import install_model_requirements
 from kipoi.utils import Slice_conv
 import config
+import numpy as np
+from kipoi.utils import cd
 
 # TODO: Implement automatic switching of backends to test on Theano model!
 
@@ -40,9 +42,27 @@ def get_extractor_cfg(model_dir):
 def get_test_kwargs(model_dir):
     return read_json_yaml(os.path.join(model_dir, 'example_files/test.json'))
 
-def get_sample_functional_model_input():
+def get_sample_functional_model_input(kind = "list"):
     import numpy as np
-    return [np.random.rand(1,280,256), np.random.rand(1,280,256)]
+    if kind == "list":
+        return [np.random.rand(1,280,256), np.random.rand(1,280,256)]
+    if kind == "dict":
+        return {"FirstInput": np.random.rand(1,280,256), "SecondInput":  np.random.rand(1,280,256)}
+
+def get_single_layer_model():
+    from keras.layers import Dense
+    from keras.models import Sequential
+    model = Sequential()
+    model.add(Dense(1, input_shape=(1,)))
+    model.layers[0].set_weights([np.array([[0.5]]), np.array([0.0])])
+    import tempfile
+    temp_weights = tempfile.mkstemp()[1]
+    temp_arch = tempfile.mkstemp()[1]
+    with open(temp_arch, "w") as ofh:
+        ofh.write(model.to_json())
+    model.save_weights(temp_weights)
+    return temp_weights, temp_arch
+
 
 def get_sample_functional_model():
     # Keras ["2.0.4", "2.1.5"]
@@ -53,8 +73,8 @@ def get_sample_functional_model():
         from keras.layers import Input, LSTM, Dense
         from keras.models import Model
         #
-        tweet_a = Input(shape=(280, 256))
-        tweet_b = Input(shape=(280, 256))
+        tweet_a = Input(shape=(280, 256), name="FirstInput")
+        tweet_b = Input(shape=(280, 256), name="SecondInput")
         #
         # This layer can take as input a matrix
         # and will return a vector of size 64
@@ -83,8 +103,8 @@ def get_sample_functional_model():
         from keras.layers import Input, LSTM, Dense, merge
         from keras.models import Model
         #
-        tweet_a = Input(shape=(280, 256))
-        tweet_b = Input(shape=(280, 256))
+        tweet_a = Input(shape=(280, 256), name="FirstInput")
+        tweet_b = Input(shape=(280, 256), name="SecondInput")
         #
         # This layer can take as input a matrix
         # and will return a vector of size 64
@@ -244,3 +264,100 @@ def test_generate_activation_output_functions():
     act_fn_nl(sample_input)
     act_fn_l = model._generate_activation_output_functions(layer="hidden", pre_nonlinearity=True)
     act_fn_l(sample_input)
+
+def test_single_layer_gradient():
+    model = kipoi.model.KerasModel(*get_single_layer_model())
+    sample_input = np.array([1])
+    grad_out = model.input_grad(sample_input, wrt_final_layer=True, avg_func="absmax")
+    assert np.all(grad_out == 0.5)
+
+
+@pytest.mark.parametrize("example", EXAMPLES_TO_RUN)
+def test_gradient_function_model(example):
+    """Test extractor
+    """
+    if example == "rbp" and sys.version_info[0] == 2:
+        pytest.skip("rbp example not supported on python 2 ")
+    #
+    example_dir = "examples/{0}".format(example)
+    # install the dependencies
+    # - TODO maybe put it implicitly in load_dataloader?
+    if INSTALL_REQ:
+        install_model_requirements(example_dir, "dir", and_dataloaders=True)
+    #
+    Dl = kipoi.get_dataloader_factory(example_dir, source="dir")
+    #
+    test_kwargs = get_test_kwargs(example_dir)
+    #
+    # install the dependencies
+    # - TODO maybe put it implicitly in load_extractor?
+    if INSTALL_REQ:
+        install_model_requirements(example_dir, source="dir")
+    #
+    # get model
+    model = kipoi.get_model(example_dir, source="dir")
+    #
+    with cd(example_dir + "/example_files"):
+        # initialize the dataloader
+        dataloader = Dl(**test_kwargs)
+        #
+        # sample a batch of data
+        it = dataloader.batch_iter()
+        batch = next(it)
+        # predict with a model
+        model.predict_on_batch(batch["inputs"])
+        model.input_grad(batch["inputs"], Slice_conv()[:, 0])
+        model.input_grad(batch["inputs"], avg_func = "sum")
+        # if example == "rbp":
+        #    model._input_grad(batch["inputs"], -1, Slice_conv()[:, 0])
+        # elif example == "extended_coda":
+        #    model._input_grad(batch["inputs"], -1, filter_func=tf.reduce_max, filter_func_kwargs={"axis": 1})
+
+
+
+
+def check_grad(input, grad):
+    if isinstance(input, list):
+        for el, grad_el in zip(input, grad):
+            assert el.shape == grad_el.shape
+    else:
+        assert input.shape == grad.shape
+
+def test__get_gradient_function():
+    import keras.backend as K
+    model = kipoi.model.KerasModel(*get_sample_functional_model())
+    with pytest.raises(Exception):# expect exception
+        grad_fn = model._get_gradient_function(use_final_layer = True)
+    sample_input=get_sample_functional_model_input()
+    grad_fn = model._get_gradient_function(use_final_layer = True, filter_slices = 0)
+    check_grad(sample_input, grad_fn(sample_input))
+    grad_fn = model._get_gradient_function(use_final_layer = False, layer = "shared_lstm", filter_slices = 0)
+    check_grad(sample_input, grad_fn(sample_input))
+    grad_fn = model._get_gradient_function(use_final_layer = True, filter_func = K.sum)
+    check_grad(sample_input, grad_fn(sample_input))
+    grad_fn = model._get_gradient_function(use_final_layer = False, layer = "shared_lstm", filter_func = K.sum)
+    check_grad(sample_input, grad_fn(sample_input))
+    # seqeuntial model:
+    model = kipoi.model.KerasModel(*get_sample_sequential_model())
+    sample_input=get_sample_sequential_model_input()
+    grad_fn = model._get_gradient_function(use_final_layer = True, filter_slices = 0)
+    check_grad(sample_input, grad_fn(sample_input))
+    grad_fn = model._get_gradient_function(use_final_layer = False, layer = 2, filter_slices = 0)
+    check_grad(sample_input, grad_fn(sample_input))
+    grad_fn = model._get_gradient_function(use_final_layer = True, filter_func = K.sum)
+    check_grad(sample_input, grad_fn(sample_input))
+    grad_fn = model._get_gradient_function(use_final_layer = False, layer = 2, filter_func = K.sum)
+    check_grad(sample_input, grad_fn(sample_input))
+
+
+def test_returned_gradient_fmt():
+    model = kipoi.model.KerasModel(*get_sample_functional_model())
+    sample_input = get_sample_functional_model_input(kind="list")
+    grad_out = model.input_grad(sample_input, wrt_final_layer = True, avg_func = "absmax")
+    assert isinstance(grad_out, type(sample_input))
+    assert len(grad_out) == len(sample_input)
+    sample_input = get_sample_functional_model_input(kind="dict")
+    grad_out = model.input_grad(sample_input, wrt_final_layer=True, avg_func="absmax")
+    assert isinstance(grad_out, type(sample_input))
+    assert len(grad_out) == len(sample_input)
+    assert all([k in grad_out for k in sample_input])
